Implement `column` from `unsafeIndex` (#57)
* Implement `column` from `unsafeIndex`

This gets rid of the temporary vector that `U.enumFromStepN`was allocating. This vector was seemingly preventing stream fusion to occur in `multiply`. According to the benchmarks from #56, on 100*100 matrices, the new `column` is twice as fast, allocates half as much memory, and `multiply` goes from allocating 17,040,080 bytes to 240,064, and is three times faster